Split management.go in multiple files
